Rename guitar chords to t_guitar_chords
